increase ms5637 conversion time at 8192 resolution
the original 17 ms delay for conversion time caused i2c abort errors when reading the ADC result under certain conditions. this was most likely occurring when the ms5637 sensor thread was able to execute immediately after the delay (which would have been rounded up to 20 ms), due to all other threads being blocked. however this also means the 20 ms delay was insufficient for the ms5637 to finish conversion, which does not match the datasheet.